+++ src/New/Entity/AttachEffectClass.h
@@ -112,6 +112,7 @@ struct AttachEffectTechnoProperties
 	double ArmorMultiplier;
 	double SpeedMultiplier;
 	double ROFMultiplier;
+	double NegativeDamageMultiplier;
 	bool Cloakable;
 	bool ForceDecloak;
 	bool DisableWeapons;
@@ -122,12 +123,15 @@ struct AttachEffectTechnoProperties
 	bool HasOnFireDiscardables;
 	bool HasRestrictedArmorMultipliers;
 	bool HasCritModifiers;
+	bool HasExtraWarheads;
+	bool HasFeedbackOrAuxWeapon;
 
 	AttachEffectTechnoProperties() :
 		FirepowerMultiplier { 1.0 }
 		, ArmorMultiplier { 1.0 }
 		, SpeedMultiplier { 1.0 }
 		, ROFMultiplier { 1.0 }
+		, NegativeDamageMultiplier { 1.0 }
 		, Cloakable { false }
 		, ForceDecloak { false }
 		, DisableWeapons { false }
@@ -138,5 +142,7 @@ struct AttachEffectTechnoProperties
 		, HasOnFireDiscardables { false }
 		, HasRestrictedArmorMultipliers { false }
 		, HasCritModifiers { false }
+		, HasExtraWarheads { false }
+		, HasFeedbackOrAuxWeapon { false }
 	{ }
 };